Store now is separate from the complicated implementations of Store
