+++ src/lib/utils.ts
@@ -1,3 +1,5 @@
+export type Pair<K, V> = { Key: K, Value: V };
+
 export function captureStack() {
   try { throw new Error(); } catch (e) { return e.stack; }
 }
